chore: api연결 오류 수정 ver1

+++ domain/src/main/java/com/weit/domain/model/journal/TravelJournalContentsInfo.kt
@@ -4,8 +4,8 @@ data class TravelJournalContentsInfo(
     val travelJournalContentId: Long,
     val content: String,
     val placeId: String,
-    val latitude: List<Long>,
-    val longitude: List<Long>,
+    val latitude: List<Double>,
+    val longitude: List<Double>,
     val travelDate: String,
     val travelJournalContentImages: List<TravelJournalContentsImagesInfo>,
 )
